demo/prompts_retrivel/requirements.txt
faiss-cpu
langchain
loguru
openai
streamlit_chat
tiktoken
tqdm
guidance
pydantic
lark
pinecone-client
pypdf